+++ workspaces/github-actions/plugins/github-actions/src/components/Cards/Cards.tsx
@@ -16,6 +16,7 @@
 
 import { useEntity } from '@backstage/plugin-catalog-react';
 import LinearProgress from '@material-ui/core/LinearProgress';
+import Tooltip from '@material-ui/core/Tooltip';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import ExternalLinkIcon from '@material-ui/icons/Launch';
@@ -63,6 +64,13 @@ const WidgetContent = (props: {
               status={lastRun.status}
               conclusion={lastRun.conclusion}
             />
+          </Box>
+        ),
+        age: (
+          <Box display="flex">
+            <Tooltip title={lastRun.statusDate ?? ''}>
+              <Box>{lastRun.statusAge}</Box>
+            </Tooltip>
           </Box>
         ),
         message: lastRun.message,