+++ workspaces/kiali/plugins/kiali/dev/index.tsx
@@ -21,7 +21,6 @@ import { createDevApp } from '@backstage/dev-utils';
 import { EntityProvider } from '@backstage/plugin-catalog-react';
 import { TestApiProvider } from '@backstage/test-utils';
 import { Grid } from '@material-ui/core';
-import { getAllThemes } from '@redhat-developer/red-hat-developer-hub-theme';
 import { EntityKialiResourcesCard, kialiPlugin } from '../src';
 import { KialiHelper } from '../src/pages/Kiali/KialiHelper';
 import { KialiNoAnnotation } from '../src/pages/Kiali/KialiNoAnnotation';
@@ -154,7 +153,6 @@ const MockKialiError = () => {
 
 createDevApp()
   .registerPlugin(kialiPlugin)
-  .addThemes(getAllThemes())
   .addPage({
     element: <KialiMock />,
     title: 'KialiPage',