+++ workspaces/todo/plugins/todo/src/alpha.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { createDevApp } from '@backstage/dev-utils';
-import { Model } from '@patternfly/react-topology';
-import { TrafficGraph } from '../src';
-import graphModel from './__data__/bookinfoGraphModel.json';
 
-createDevApp()
-  .registerPlugin()
-  .addPage({
-    element: <TrafficGraph model={graphModel as Model} />,
-    title: 'TrafficGraph',
-    path: '/kiali/trafficGraph',
-  });
+export * from './alpha/index';
+export { default } from './alpha/index';