+++ workspaces/kiali/plugins/kiali/CHANGELOG.md
@@ -1,5 +1,13 @@
 ### Dependencies
 
+## 1.39.2
+
+### Patch Changes
+
+- 1df9013: remove product theme from dev dependencies and dev app
+- Updated dependencies [1df9013]
+  - @backstage-community/plugin-kiali-react@0.1.2
+
 ## 1.39.1
 
 ### Patch Changes