+++ workspaces/sonarqube/packages/app-next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # app-next
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [6d6b976]
+  - @backstage-community/plugin-sonarqube@0.17.0
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [1eda234]
+  - @backstage-community/plugin-sonarqube@0.16.2
+
 ## 0.0.21
 
 ### Patch Changes